Added new led-light status system to the main-map-view's statusbar: 1. These led-lights will flash slowly if something 'interesting' is going on 2. There will be a tooltip if mouse cursor is over led-light 3. Now there is 3 led-light on left edge of statusbar, which report different things: a) querydata based status 'channel' that tells if files are copied from server, combined-data is being made or sounding-index data is being calculated b) Wms activity led-light will report if getcapability request are being made, sadly layer-image being fetched couldn't been added here to an activity c) general problem 'channel', which now reports only if hard-drives are being filled 4. In Settings dialog is option if this led-light status system is used or not 5. Had to re-add different color led-light icons to the SmartMet project in order to make them work correctly with 16x16 size led images 6. Created new CStatusBarResize class that can handle new led-lights - Also it's OnSize method can handle better the status text pane section's width - Default system has odd behaviour where that section's minimum width is 1/4 of monitor's width (this messed up things in certain situations) 7. Added new classes that handle these new led-light functionalities like: - NFmiLedLightStatus, NFmiLedLightStatusSystem

+++ src/mfcutils/DialogEx.cpp
@@ -455,7 +455,7 @@ BOOL CDialogEx2::OnToolTipText(UINT, NMHDR* pNMHDR, LRESULT* pResult)
 	TCHAR szFullText[256];
 	CString strTipText;
 	CString strPromtpText;
-	UINT nID = pNMHDR->idFrom;
+	UINT nID = (UINT)pNMHDR->idFrom;
 
 	if (pNMHDR->code == TTN_NEEDTEXTA && (pTTTA->uFlags & TTF_IDISHWND) ||
 		pNMHDR->code == TTN_NEEDTEXTW && (pTTTW->uFlags & TTF_IDISHWND)) {
